requirements.txt
streamlit
groq
python-dotenv
joblib
scikit-learn
watchdog